add package

+++ database/migrations/2024_01_15_122746_create_packages_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->smallInteger('limit');
-            $table->integer('price');
+            $table->smallInteger('limit'); // limit days
+            $table->double('price');
             $table->timestamps();
         });
     }
